pyproject.toml
[tool.poetry]
name = "odin-messaging-bot"
version = "0.1.13"
description = "A NATS/STAN Bot to publish/subscribe messages."
authors = ["adolfrodeno <amvillalobos@uc.cl>"]

[tool.poetry.dependencies]
python = "^3.8"
nest-asyncio = "^1.5.1"
asyncio-nats-streaming = "^0.4.0"
aiohttp = "^3.3.1"
python-dotenv = "^0.19.1"
pydantic = "^1.8.2"
ujson = "^5.1.0"
asyncio-nats-client = "0.11.4"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
autopep8 = "^1.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
